add map to homepage
