+++ src/app/home/home.component.html
@@ -41,6 +41,47 @@
       </td>
     </div>
   </div>
+  <div class="row main-row">
+    <div class="col m12 s12">
+      <div class="about-us-content" *ngIf="developers">
+        <div class="row">
+          <div>
+            <agm-map
+                    [latitude]="lat "
+                    [longitude]="lng "
+                    [zoom]="zoom ">
+
+              <agm-marker
+                  *ngFor="let d of developers "
+                  [latitude]="d.latitude "
+                  [longitude]="d.longitude "
+                  (mouseOver)="handleMarkerMouseOver($event) "
+                  >
+
+                <agm-info-window>
+                  <div class="info-window ">
+                    <a class="info-container " [routerLink]="[ '/user/view',d.id] ">
+                      <div class="info-avatar ">
+                        <img [src]="d.avatarUrl " onError="this.src='../../assets/default_avatar.png' ">
+                      </div>
+                      <div class="user-name ">
+                        {{d.userName}}
+                      </div>
+                      <div class="user-location ">
+                        <span *ngIf="d.state ">{{d.state}}</span>
+                        <span *ngIf="d.state && d.country ">,</span>
+                        <span *ngIf="d.country ">{{getCountryName(d.country)}}</span>
+                      </div>
+                    </a>
+                  </div>
+                </agm-info-window>
+              </agm-marker>
+            </agm-map>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
   <div class="row github_cta">
     <div class="col s8 offset-s2">
       <p>
